Multiple scaling in a disjoint manner using negative items too

disj_nci is an internal function and returns a list of data frames. It is a loop which runs through the following steps: (1) Set up a (first) scale. (2) find the two items with the highest positive correlation in the data set. (3) If the absolute value of this correlation is greater than the pre-specified lower bound (mrit_min), add up the two items to build the core of the emerging scale. (4) While the absolute value of the correlation between the sum score and a remaining items in the data frame is greater than mrit_min, flavor the scale with the appropriate item. If necessary, reverse the item. (5) If there are at least two leftovers in the data frame, that meet the inclusion criterion, start over again.

disj_nci(df, mrit_min, sclvals, use)

Arguments

df
a data frame object.

mrit_min
a numerical constant to specify the marginal corrected item total correlation. The value must be in the range of 0-1.

sclvals
a numerical vector of length 2 indicating the start and endpoint of a scale.

use
an optional string to specify how missing values will enter the analysis. See use in cor for details.

Details

disj_nci is an internal function which extends the scope of disj_pci. disj_nci allows the presence of negative items (i.e., a negative correlation between item and sum score).

use clarifies how to set up a correlation matrix in the presence of missing values. In a typical scaling process this happens at least twice. First, when determining the core items (the two items in the correlation matrix with the highest linear relationship). Second, when an item is proposed for an emerging scale.

References

Müller-Schneider, T. (2001). Multiple Skalierung nach dem Kristallisationsprinzip / Multiple Scaling According to the Principle of Crystallization. Zeitschrift für Soziologie, 30(4), 305-315. https://doi.org/10.1515/zfsoz-2001-0404 @importFrom stats cor