Multiple scaling in a disjoint manner

disjoint returns a multiple, disjoint scaled version of the specified data frame. This so called msdf sets up the building block for further analysis with overlap (type ?overlap).

disjoint(
  df,
  mrit_min = 0.3,
  negative_too = FALSE,
  sclvals = NULL,
  use = "pairwise.complete.obs"
)

Arguments

df
a data frame (with more than two items and unique, non-NA column names).

mrit_min
a numeric constant of length 1 to specify the marginal corrected item-total correlation. Its value is in the range of 0-1. The default is set to .3.

negative_too
a logical constant indicating whether reversed items are included in the analysis. The default is set to FALSE.

sclvals
a numeric vector of length 2 indicating the start- and endpoint of a scale. Use something like c(min,max).

use
an optional string to specify how missing values enter the analysis. See use in cor for details. The default is set to pairwise.complete.obs.

Details

use clarifies how to set up a correlation matrix in the presence of missing values. In a typical scaling process this happens at least twice. First, when determining the core items (the two items in the correlation matrix with the highest linear relationship). Second, when an item is proposed for an emerging scale.

References

Müller-Schneider, T. (2001). Multiple Skalierung nach dem Kristallisationsprinzip / Multiple Scaling According to the Principle of Crystallization. Zeitschrift für Soziologie, 30(4), 305-315. https://doi.org/10.1515/zfsoz-2001-0404

Examples

# Use only positive correlations
disjoint(mtcars, mrit_min = .4)

#> $scl_1
#>           mrit rbar alpha
#> disp, cyl 0.90 0.90  0.95
#> wt        0.89 0.86  0.95
#> hp        0.79 0.81  0.94
#> carb      0.55 0.70  0.92
#> 
#> $scl_2
#>          mrit rbar alpha
#> gear, am 0.79 0.79  0.89
#> drat     0.74 0.74  0.89
#> mpg      0.63 0.66  0.89
#> vs       0.63 0.54  0.86
#> 

# Include negative correlations
disjoint(mtcars, mrit_min = .4, negative_too = TRUE, sclvals = c(1,7))

#> 
#> disjoint() reversed the following item(s):
#> - mpg
#> - vs
#> - drat
#> - qsec
#> - am
#> - gear

#> $scl_1
#>           mrit rbar alpha
#> disp, cyl 0.90 0.90  0.95
#> wt        0.89 0.86  0.95
#> mpg       0.85 0.86  0.96
#> hp        0.80 0.82  0.96
#> vs        0.75 0.78  0.95
#> drat      0.65 0.73  0.95
#> qsec      0.56 0.66  0.94
#> carb      0.55 0.62  0.94
#> am        0.49 0.57  0.93
#> gear      0.42 0.53  0.93
#> 

# Change the treatment of missing values
disjoint(mtcars, mrit_min = .4, use = "all.obs")

#> $scl_1
#>           mrit rbar alpha
#> disp, cyl 0.90 0.90  0.95
#> wt        0.89 0.86  0.95
#> hp        0.79 0.81  0.94
#> carb      0.55 0.70  0.92
#> 
#> $scl_2
#>          mrit rbar alpha
#> gear, am 0.79 0.79  0.89
#> drat     0.74 0.74  0.89
#> mpg      0.63 0.66  0.89
#> vs       0.63 0.54  0.86
#>