German General Social Survey's Trust Items

A subset of the German General Social Survey (ALLBUS) 2018 containing its trust items (F018).

trust

Format

Participants were asked about their trust in public institutions and organizations. `1` means they have absolutely no trust at all, whereas `7` represents a great deal of trust. The data frame has 3477 rows and 13 variables:

healserv

Health service

fccourt

German constitutional court

bundtag

German Parliament

munadmin

Municipal administration

judsyst

Judicial system

tv

Television

newsppr

Newspapers

uni

Universities and other institutes of higher education

fedgovt

German government

police

Police

polpati

Political parties

eucomisn

European Commission

eupalmnt

European Parliament

Details

Please note that in comparison to the original data set participant order has been randomly rearranged to further ensure anonymity. This might lead to differences when trying to reproduce the results of the analysis with the original data set (mentioned below).

References

GESIS - Leibniz-Institut für Sozialwissenschaften (2019). German General Social Survey - ALLBUS 2018. GESIS Datenarchiv, Köln. ZA5272 Datenfile Version 1.0.0, https://doi.org/10.4232/1.13325.

Examples

# Use trust with disjoint
disjoint(trust, mrit_min = .4)

#> $scl_1
#>                    mrit rbar alpha
#> eupalmnt, eucomisn 0.92 0.92  0.96
#> polpati            0.64 0.72  0.89
#> fedgovt            0.68 0.67  0.89
#> bundtag            0.70 0.64  0.90
#> judsyst            0.58 0.59  0.90
#> fccourt            0.58 0.55  0.90
#> police             0.52 0.52  0.90
#> newsppr            0.52 0.49  0.90
#> tv                 0.52 0.47  0.90
#> uni                0.49 0.45  0.90
#> munadmin           0.47 0.43  0.90
#> healserv           0.42 0.40  0.90
#>