A set of utility functions

A set of utility functions

nme_msdf(x)

calc_rbar(scl, use)

calc_alpha(scl, use)

calc_mrit(scl, use)

extr_core(scl)

extr_core_nms(scl)

extreb_itms(df, itm_nms)

rvrs_var(var, sclvals)

rvrs_note(msg, applicant)

Arguments

x
a multiple scaled data frame

scl
a scale within a multiple scaled data frame

use
an optional string indicating how to deal with missing values, See use in cor for details.

df
a data frame object

itm_nms
the names of an item from a scale

var
a variable or item (often a column from a data frame)

sclvals
the start and end point of a scale (specify: c(sp,ep))

msg
a reverse message sent from either disjoint or overlap

applicant
the function which wants to leave messages disjoint or overlap)

Details

nme_msdf() designates the components of a msdf (e.g., `scl_1`).

calc_rbar() calculates the average correlation between items from a gradually emerging scale.

calc_alpha() calculates Cronbach's alpha of a given set of items from the gradually emerging scale.

calc_mrit() calculates the corrected item-total correlation of an item with the sum score of all the other variables of the gradually emerging scale (aka. marginal corrected item-total correlation). Note, that this procedure uses a part-whole correction. Thus, the item itself is excluded in the calculation process.

extreb_itms() build up the counterpart of a fragment from the given item names. Therefore, the counterpart includes all variables that are not part of a fragment but which are mentioned in the specified data set.

extr_core() extracts all pairs of core items from a fragment.

extr_core_nms() extracts the names of every core items in a given fragment.

rvrs_var() reverses an item using the specified scaling values. It handles the following types of scales:

...-3 -2 -1 0 1 2 3..., e.g., sclvals = c(-3, 3)

0 1 2 3 4 5 6..., e.g., sclvals = c(0, 7)

1 2 3 4 5 6 7..., e.g., sclcals = c(1, 7)

rvrs_note() gets the full report of reversed variables and reports a unique list of them.