elisr's quadriga

The four workhorses inside elisr's user functions disjoint() and overlap().

disj_pci() is a loop which runs through the following steps: (1) Set up a (first) scale. (2) Find the two items with the highest positive correlation in the data set. (3) If the absolute value of this correlation is greater than the pre-specified lower bound (mrit_min), add up the two items to build the core of the emerging scale. (4) As long as the value of the correlation between the sum-score and a remaining item in the data frame is greater than mrit_min, flavor the scale with the appropriate item. (5) If there are at least two leftovers in the data frame that meet the inclusion criterion, start over again.

disj_nci() is almost identical to disj_pci(), though step (4) varies slightly from above. To take negative correlations into account, disj_nci() flavors the scale with appropriate item as long as the absolute value of the correlation between the sum-score and a remaining items in the data frame is greater than mrit_min.

ovlp_pci() takes a disjointedly built scale fragment and tries to extend it with those items in the data set, which are not yet built into the fragment (aka., its counterpart). Because ovlp_pci() does this for every disjointedly built scale fragment it is a multiple one-dimensional extension of disj_pci().

The only difference to ovlp_pci() is that ovlp_nci() can handle reversed items. The extension algorithm remains almost the same; ovlp_nci() flavors each scale fragment with appropriate items from its counterpart as long as the absolute value of the correlation between the sum-score and a remaining item is greater than mrit_min. Thus, it is a multiple one-dimensional extension of disj_nci():

disj_pci(df, mrit_min, use)

disj_nci(df, mrit_min, sclvals, use)

ovlp_pci(msdf, mrit_min, overlap_with, use)

ovlp_nci(msdf, mrit_min, overlap_with, sclvals, use)

Arguments

df
a data frame object.

mrit_min
a numerical constant to specify the marginal corrected item total correlation. The value must be in the range of 0-1.

use
an optional string to specify how missing values will enter the analysis. See use in cor for details.

sclvals
a numerical vector of length 2 indicating the start- and endpoint of a scale.

msdf
a multiple scaled data frame (built with disjoint()).

overlap_with
a string telling overlap() the set of items for the extension. To build up on all variables of a fragment use fragment, for the core-only option type core. The default is set to "fragment".

Value

disj_pci() and disj_nci() both return a list of data frames which result from applying the above-mentioned algorithm.

ovlp_pci() and ovlp_nci() often return an extended a list of data frames.

Details

All functions are internal functions.

The use argument specifies how to set up a correlation matrix in the presence of missing values. In a typical scaling process this happens at least twice. First, when determining the core items (the two items in the correlation matrix with the highest linear relationship). Second, when an item is proposed for an emerging scale.

Note that all functions use cor's default method pearson.

References

Müller-Schneider, T. (2001). Multiple Skalierung nach dem Kristallisationsprinzip / Multiple Scaling According to the Principle of Crystallization. Zeitschrift für Soziologie, 30(4), 305-315. https://doi.org/10.1515/zfsoz-2001-0404