A set of checks for the appropriateness of inputs

A set of checks for the appropriateness of inputs

check_df(x)

check_sclvals(x)

compare_sclvals(x, x_attr)

check_mrit(x)

check_ovlp(x)

check_msdf(x)

check_neg(x)

check_comp(x, mrit_min, use)

Arguments

x
some arbitrary input to be checked

x_attr
a numeric vector of length 2 indicating the start- and endpoint of a scale.

mrit_min
a numeric constant of length 1 to specify the marginal corrected item-total correlation.

use
an optional string to specify how missing values enter the analysis. See use in cor for details.

Details

check_df(x) guarantees that x is a data frame. In addition, it verifies that x has less than two variables (a single item can't build a core), x has column names (used to pre-build scls in the overlapping process), if the column names are unique and not of type NA.

check_sclvals(x) checks whether x is a two element vector (Note: implicitly assures that it is not NULL, since NULL is a logical constant of length `0`.). Integers are coerced to be of type double. Additionally the function makes sure that the first value is not greater than the second.

compare_sclvals(x) ensures that the sclvals set with overlap() are equal to those set with disjoint().

check_rit(x) guarantees that the input is a double vector of length one. Moreover, it secures that the lower bound is unique and ranges between `0` and `1`.

check_ovlp(x) safeguards thatx, is a character vector of length `1`. Within disjoint() and overlap()) switch() validates the input string itself, throwing an error message when the given character doesn't match any available option.

check_msdf(x) guards against inputs that are not of type `msdf`.

check_neg(x) verifies that the (negative_too input is a logical constant of length 1 and not a missing value (this is necessary because objects of type NA are logical constants of length 1, too).

check_comp examines the correlation matrix, cor(df). It complains if no correlation is greater than the specified mrit_min.